Mkh -> main / 작업 옮기기

+++ Assets/Project/Programmer/NSJ/Script/AdditionalEffect/BlueChip/PlayerAdditional/Script/PlayerAdditional.cs
@@ -17,8 +17,8 @@ public class PlayerAdditional : AdditionalEffect
         Player = player;
     }
 
-    protected int GetPlayerAttackPower(int attackPower)
+    protected float GetPlayerAttackPower(int attackPower)
     {
-        return (Model.AttackPower - (int)Model.Data.EquipStatus.Damage) + attackPower;
+        return (Model.AttackPower - Model.Data.EquipStatus.Damage) + attackPower;
     }
 }
